Make some required changes
