+++ spec/features/post_index_spec.rb
@@ -1,13 +1,13 @@
 require 'rails_helper'
-# rubocop:disable Metrics/BlockLength
+
 RSpec.feature 'Posts', type: :feature do
-  before(:all) do
+  before do
     @user1 = User.create(name: 'Patrick Mboma', photo: 'http://twitter.com', bio: 'test for User')
     @user2 = User.create(name: 'Paul Biya', photo: 'http://twitter.com', bio: 'test for User')
     @post1 = Post.create(title: 'Rspec test 1', text: 'rspec test for post', user: @user1)
     @comment1 = Comment.create(user: @user1, post: @post1, text: 'test for User')
     @comment2 = Comment.create(user: @user2, post: @post1, text: 'test for User2')
-    Like.create(user: @user1, post: @post1)
+    @like = Like.create(user: @user1, post: @post1)
   end
 
   describe 'User post index page' do
@@ -62,45 +62,4 @@ RSpec.feature 'Posts', type: :feature do
       expect(page).to have_current_path(user_post_path(@user1, @post1))
     end
   end
-
-  describe 'User post show page' do
-    scenario 'Should show the post title ' do
-      visit user_post_path(@user1, @post1)
-      expect(page).to have_content(@post1.title)
-    end
-
-    scenario 'Should show the username ' do
-      visit user_post_path(@user1, @post1)
-      expect(page).to have_content(@user1.name)
-    end
-
-    scenario 'Should show the number of comment ' do
-      visit user_post_path(@user1, @post1)
-      expect(page).to have_content("Comments: #{@post1.comments.count}")
-    end
-
-    scenario 'Should show the number of likes ' do
-      visit user_post_path(@user1, @post1)
-      expect(page).to have_content("Likes: #{@post1.likes.count}")
-    end
-
-    scenario 'Should show the post body ' do
-      visit user_post_path(@user1, @post1)
-      expect(page).to have_content(@post1.text)
-    end
-
-    scenario 'Should show the post commentor name ' do
-      visit user_post_path(@user1, @post1)
-      expect(page).to have_content(@comment1.user.name)
-    end
-
-    scenario 'Should show the comment each comment user left ' do
-      @user = User.first
-      @post = Post.first
-      visit user_post_path(@user, @post)
-      expect(page).to have_text(@post.comments[0].text)
-      expect(page).to have_text(@post.comments[1].text)
-    end
-  end
 end
-# rubocop:enable Metrics/BlockLength
